requirements.txt
python_gantt==0.6.0
Django==2.0.1
django-model-utils==3.1.1
virtualenv==15.1.0
virtualenvwrapper-win==1.2.5
Pillow==5.0.0
psycopg2==2.7.3.2
django-crispy-forms==1.7.0
